+++ core/src/util/mod.rs
@@ -1,6 +1,9 @@
+mod abort_on_drop;
 pub mod db;
 #[cfg(debug_assertions)]
 pub mod debug_initializer;
 pub mod error;
 pub mod migrator;
 pub mod seeder;
+
+pub use abort_on_drop::*;
